% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-processing.R
\name{acf.adj}
\alias{acf.adj}
\title{Plots estimates of autocorrelation generated from autocorr.unc()}
\usage{
acf.adj(autocorr, plot.title = "", lag.max = NULL, show.plot = TRUE)
}
\arguments{
\item{\code{autocorr}}{\emph{list.}      Output of autocorr.unc(y) (see below), where y is a vector of the relevant data}

\item{\code{plot.title}}{\emph{character.}             Title of the desired ACF plot}

\item{\code{lag.max}}{\emph{Numeric.}               Maximum lag to be shown in ACF plot}

\item{\code{show.plot}}{\emph{logical.}               Show plot in Rstudio window or not}
}
\value{
\code{fplot}         \emph{ggplot}                Plot object of ACF plot
}
\description{
Similar to R default acf() function, can also plot arbitrary bound lines. Useful when plotting the cut off lag proposed by Zhang in Calculation of the uncertainty of the mean of autocorrelated measurements (2006).
}
\examples{
# Generate a set of 1000 numbers from the standard normal distribution
y <- rnorm(1000)

# Calculate the autocorrelation and adjusted uncertainty for y
autocorr <- autocorr.unc(y)

# Plot acf plot with adjusted bounds defined by autocorr.unc
fig <- acf.adj(autocorr)

}
\references{
Xhang NF (2006) Calculation of the uncertainty of the mean of autocorrelated measurements. \emph{Metrologia} 43(4):S276. URL http://stacks.iop.org/0026-1394/43/i=4/a=S15.
}
