\name{mouth2ear-package}
\alias{mouth2ear-package}
\alias{mouth2ear}
\docType{package}
\title{
\packageTitle{mouth2ear}
}
\description{
\packageDescription{mouth2ear}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mouth2ear}
\packageIndices{mouth2ear}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mouth2ear}

Maintainer: \packageMaintainer{mouth2ear}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
