% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-processing.R
\name{process.sessions}
\alias{process.sessions}
\title{process.sessions perform uncertainty analysis and report results for list of M2E latency tests}
\usage{
process.sessions(all.setups, show.lags = F)
}
\arguments{
\item{\code{show.lags}}{\emph{logical.}              Informs whether or not to print the maximum lag for which significant autocorrelation is present for each session of each test}

\item{\code{all.setups}}{\emph{list.}                  List containing information for all tests. Further detailed below.

  \code{name}          \emph{character.}             Name of test setup (i.e. Single Location Cabled Tests)

  \code{path}          \emph{character.}             Path where csv files containing delay values for all tests of test setup contained. Expects each test in individual folder containing a csv file for each session.

  \code{tests}         \emph{character vector.}      Vector of names of the specfic tests performed in the test setup (i.e. c("Device X Direct Mode", "Device X Trunked Mode", "Device Y Direct Mode")). Must match the folder names that contain the session csv files for that test.

  \code{thinning}      \emph{numeric vector.}        Vector of degree to which the sessions of each test must be thinned to remove significant autocorrelation. For example given Device X Direct Mode must be thinned 
                                      by using every third data point to eliminate significant autocorrelation, Device X Trunked must be thinned by using every fifth data point, and Device Y Direct 
                                      Mode must be thinned by using every second data point then thinning would be c(3,5,2).}
}
\value{
A list containing the following components:

\code{thinned.data}  \emph{list.}                  List with thinned data vectors for each session of each test of each test setup

\code{gum.data}      \emph{list.}                  List with GUM output for each session of each test of each test setup

\code{df}            \emph{data.frame.}           Dataframe with GUM() output for all tests

\code{autocorr.data} \emph{list.}                  List with autocorr.unc() output for each session of each test of each setup
}
\description{
process.sessions(all.setups) performs uncertainty analysis on the input test data in accordance with the Guide to the expression of uncertainty of measurement, GUM. Relies on the 
  functions GUM and GUM.validate from the metRology package.
}
\details{
process.sessions(all.setups, show.lags = T) prints the maximum lag for which significant autocorrelation is present for each session of each test.
}
