% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-processing.R
\name{acf.adj}
\alias{acf.adj}
\title{Plots estimates of autocorrelation generated from autocorr.unc()}
\usage{
acf.adj(autocorr, plot.title = "", lag.max = NULL, show.plot = TRUE)
}
\description{
Similar to R default acf() function, can also plot arbitrary bound lines. Useful when plotting the cut off lag proposed by Zhang in Calculation of the uncertainty of the mean of autocorrelated measurements (2006).
}
\details{
-------------------------Inputs-------------------------------------------
  NAME          TYPE                  DESCRIPTION
  autocorr      list                  Output of autocorr.unc(y) (see below), where y is a vector of the relevant data

plot.title    character             Title of the desired ACF plot

lag.max       Numeric               Maximum lag to be shown in ACF plot

show.plot     Boolean               Show plot in Rstudio window or not

-------------------------Outputs-------------------------------------------
  NAME          TYPE                  DESCRIPTION
  fplot         ggplot                Plot object of ACF plot
}
