% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-processing.R
\name{autocorr.unc}
\alias{autocorr.unc}
\title{Determine if data autocorrelated and calculate corrected uncertainty autocorrelated data}
\usage{
autocorr.unc(y)
}
\arguments{
\item{\code{y}}{\emph{numeric vector}        Vector of data on which to calculate sample autocorrelation and uncertainty}
}
\value{
A list with the following elements:

\code{u}             \emph{numeric.}               Uncertainty corrected for autocorrelation if significant autocorrelation detected

\code{lag}           \emph{numeric.}               Maximum lag for which significant autocorrelation present (i.e. element k and element k+lag have significant autocorrelation)

\code{rho}           \emph{numeric vector.}        Sample autocorrelation

\code{sigma}         \emph{numeric Vector.}        Sample autocorrelation variance estimator

\code{r}             \emph{numeric.}               Ratio between uncertainty of corrected uncertainty versus uncorrected uncertainty
}
\description{
Determine if data autocorrelated and calculate corrected uncertainty autocorrelated data
}
\references{
Xhang NF (2006) Calculation of the uncertainty of the mean of autocorrelated measurements. \emph{Metrologia} 43(4):S276. URL http://stacks.iop.org/0026-1394/43/i=4/a=S15.
}
