\name{mouth2ear-package}
\alias{mouth2ear-package}
\alias{mouth2ear}
\docType{package}
\title{
\packageTitle{mouth2ear}
}
\description{
\packageDescription{mouth2ear}
}
\details{

\packageDESCRIPTION{mouth2ear}
\packageIndices{mouth2ear}
The mouth2ear package includes functions for:
\enumerate{
  \item Calculating M2E latency values for tests taken across multiple sessions. Can further thin data to eliminate significant autocorrelation in data.
  \item Creating acf plots and identifying the effect of autocorrelation on measurement uncertainty.
}
}
\author{
\packageAuthor{mouth2ear}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
